feat: added partial component struct and inertia request type
PartialComponent is a struct that holds the component name, the partial data and data exceptions.

InertiaRequestType is an enum that holds information about the request: if it is standard, then all props is sent. If it's partial, only props with key in `only` is sent, also if there is no only, data that are on `except` won't be sent.